ChannelInfo: Reset channel_info_count in callback; fix Reserved set
Right now we're getting called twice when two adapters
are installed (probably a problem with using NLM_F_DUMP),
which causes channel_info_count to be incremented again
from the second handler call.

Also fix our resetting of the Reserved set, which was reduced
to two in the latest CACE API to make room for the Flag parameter.